review-warnings: fix runtime/src (rest) warnings
Address the runtime/src warnings outside mm/ (lib, kernel_ffi, log,
msg_router, sched, component, loader, inference) from
REVIEW_WARNINGS.md.

Code fixes:
- lib.rs panic handler: added `static AtomicBool PANICKING` recursion
  guard. On second entry (e.g. uart_printf wedged in lwIP TX, timer
  IRQ re-enters the runtime), hard-loops on spin_loop instead of
  recursing infinitely. First panic line still reaches UART before
  the second entry.
- kernel_ffi.rs: renamed the Rust binding `panic` to `slm_panic`
  with #[link_name = "panic"] so the C-side symbol stays unchanged.
  A future `use kernel_ffi::*` no longer shadows Rust's panic!
  macro lexically. Updated lib.rs caller.
- sched/heterogeneous.rs big_little: switched to
  saturating_add(...).min(MAX_CORES) so big_cores + little_cores >
  255 produces a sensible MAX_CORES topology instead of wrap.
- loader/protobuf.rs: switched the cursor advance to
  checked_add().ok_or(LengthOverflow). usize wrap is unreachable in
  practice but the guard is cheap.
- loader/onnx_parser.rs: switched u64 → i64 dim cast to
  i64::try_from(v).map_err(|_| LengthOverflow). A varint dim above
  i64::MAX now fails parsing instead of producing a negative dim.
- loader/registry.rs: replaced direct src indexing in the FP16→FP32
  loop with src.get(...). Truncated FP16 raw_data now produces
  CorruptedData instead of a panic.
- inference/ops.rs matmul_int8: promoted accumulator to i64. With
  K up to ~2^47 the accumulator can't overflow; the previous i32
  could overflow at K = ~32768 with extreme zero-points.
- inference/engine.rs exec_reshape: explicitly reject negative dims
  other than -1 (return ShapeOverflow). Also reject positive dims
  that don't fit in u32. Previously both silently truncated.
- inference/tensor.rs: removed `unsafe impl Sync for Tensor`. Sync
  is unsound because data_mut(&self) lets a shared &Tensor produce
  a writable pointer. Higher-level locks already serialize accesses;
  the comment marks the constraint so a future maintainer doesn't
  add Sync without first hiding data_mut behind &mut self.

Most other runtime/src warnings were either already correct
(reviewer agreed in the warning text), defensive-helper
recommendations covered by documented contracts, or refactoring
tied to broader work. Full disposition + reasoning in
REVIEW_WARNINGS.md.

Build verified clean on QEMU_VIRT (ARM64) and X86_64.

+++ runtime/src/inference/ops.rs
@@ -106,11 +106,18 @@ unsafe fn matmul_int8(
 
     for i in 0..m {
         for j in 0..n {
-            let mut acc: i32 = 0;
+            // i64 accumulator: a_val/b_val are in [-256, 255] after
+            // zero-point subtraction, so a_val * b_val is in
+            // [-65536, 65536]. With K = 32768 the running sum can
+            // reach ~2^31 and overflow i32. i64 has headroom for
+            // any realistic K value (would need ~2^47 K to overflow
+            // i64, far past any plausible weight matrix). The cast
+            // back to f32 truncates harmlessly.
+            let mut acc: i64 = 0;
             for kk in 0..k {
                 let a_val = (*ap.add(i * k + kk) as i32) - zp_a;
                 let b_val = (*bp.add(kk * n + j) as i32) - zp_b;
-                acc += a_val * b_val;
+                acc += (a_val * b_val) as i64;
             }
             *cp.add(i * n + j) = (acc as f32) * scale;
         }
